+++ dbms/src/Flash/Coprocessor/FineGrainedShuffle.h
@@ -14,7 +14,6 @@
 
 #pragma once
 
-#include <Flash/Coprocessor/DAGContext.h>
 #include <common/types.h>
 #include <tipb/executor.pb.h>
