+++ dbms/src/Flash/Coprocessor/RemoteRequest.cpp
@@ -14,6 +14,7 @@
 
 #include <Common/FmtUtils.h>
 #include <Flash/Coprocessor/ChunkCodec.h>
+#include <Flash/Coprocessor/DAGContext.h>
 #include <Flash/Coprocessor/RemoteRequest.h>
 #include <Storages/MutableSupport.h>
 #include <common/logger_useful.h>
