+++ dbms/src/Flash/Statistics/ExecutorStatistics.h
@@ -31,6 +31,8 @@
 
 namespace DB
 {
+class DAGContext;
+
 template <typename ExecutorImpl>
 class ExecutorStatistics : public ExecutorStatisticsBase
 {
