+++ dbms/src/Flash/tests/gtest_storage_disaggregated.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include <Core/TiFlashDisaggregatedMode.h>
+#include <Flash/Coprocessor/DAGContext.h>
 #include <Storages/StorageDisaggregated.h>
 #include <TestUtils/ExecutorTestUtils.h>
 #include <TestUtils/mockExecutor.h>
